// src/generated/GqlTypes.ts
export namespace GqlTypes {
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
};



export type Block_Height = {
  hash?: Maybe<Scalars['Bytes']>;
  number?: Maybe<Scalars['Int']>;
};


export enum OrderDirection {
  Asc = 'asc',
  Desc = 'desc'
}

export type Query = {
  __typename?: 'Query';
  swap?: Maybe<Swap>;
  swaps: Array<Swap>;
  tradingPair?: Maybe<TradingPair>;
  tradingPairs: Array<TradingPair>;
  token?: Maybe<Token>;
  tokens: Array<Token>;
  user?: Maybe<User>;
  users: Array<User>;
  referralFee?: Maybe<ReferralFee>;
  referralFees: Array<ReferralFee>;
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
};
// ...
export type ReferralFee = {
  __typename?: 'ReferralFee';
  id: Scalars['ID'];
  receiver: Scalars['Bytes'];
  amount: Scalars['BigInt'];
  token: Token;
  contract?: Maybe<Scalars['Bytes']>;
  timestamp: Scalars['BigInt'];
};
// ...
export type Swap = {
  __typename?: 'Swap';
  id: Scalars['ID'];
  contract: Scalars['Bytes'];
  pair: TradingPair;
  referrer: Scalars['Bytes'];
  referrerFee: Scalars['BigInt'];
  fromAmount: Scalars['BigInt'];
  toAmount: Scalars['BigInt'];
  sender: User;
  blockNumber: Scalars['BigInt'];
  timestamp: Scalars['BigInt'];
  gasUsed: Scalars['BigInt'];
  gasPrice: Scalars['BigInt'];
};
// ...
export type Token = {
  __typename?: 'Token';
  id: Scalars['ID'];
  symbol: Scalars['String'];
  name: Scalars['String'];
  decimals: Scalars['Int'];
  tradeVolume: Scalars['BigDecimal'];
  tradeCount: Scalars['BigInt'];
};
// ...
export type TradingPair = {
  __typename?: 'TradingPair';
  id: Scalars['ID'];
  fromToken: Token;
  toToken: Token;
  tradeVolume: Scalars['BigDecimal'];
  tradeCount: Scalars['BigInt'];
};
// ...
export type User = {
  __typename?: 'User';
  id: Scalars['ID'];
  trades?: Maybe<Array<Swap>>;
  tradeCount: Scalars['BigInt'];
};
// ...
export type _Meta_ = {
  __typename?: '_Meta_';
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars['String'];
};
// ...
}